README.txt
In this project, I calculate how many households are in each municipality in Ontario. Data sources are not publicly available, but the method for calculating these sources can be found within the 'households calculation' code. Most municipal data sets include single-tier, lower-tier and upper-tier municipalities. Therefore, some conditional statements were needed to avoid double counting households and determine the true value for each municipality. 

To understand how municipalities in Ontario are structured and to see a list of the 444 municipalities in Ontario click here:
https://www.ontario.ca/page/list-ontario-municipalities#section-1